Rename pkg/dynamic-instrumentation to pkg/dynamicinstrumentation
