+++ pkg/dynamicinstrumentation/module.go
